Renamed test.py to pong.py
